requirements.txt
openai==0.28.1
python-dotenv==1.0.0
PyYAML==6.0.1
colorama==0.4.6
pytest==7.4.2
